Fix tests not mocking properly
The issue was caused by there being no default constructor, so to fix the issue we just throw in some random data (basically we mock the real constructor).

The second issue was not having an interface, however the code is too unholy & cursed to actually be able of implementing an interface now (so we just do a similar approach as initially with the mocking, and pray it all goes okay).